Untar Panther file and read through it

+++ orthopairs/src/main/java/org/reactome/release/orthopairs/Main.java
@@ -21,8 +21,7 @@ public class Main
 
 
     
-    public static void main( String[] args ) throws FileNotFoundException, IOException, ParseException
-    {
+    public static void main( String[] args ) throws FileNotFoundException, IOException, ParseException {
         String pathToConfig = "";
         if (args.length > 0) {
             pathToConfig = args[0];
@@ -36,40 +35,41 @@ public class Main
         String pathToSpeciesConfig = props.get("pathToSpeciesConfig").toString();
         String pathToXMLQuery = props.get("pathToXMLQuery").toString();
         URL pantherFileURL = new URL(props.get("pantherFileURL").toString());
-        String pantherFilename = releaseNumber + "_" + props.get("pantherFilename").toString();
-        File pantherFile = new File(pantherFilename);
+        String pantherFilename = props.get("pantherFilename").toString();
+        String pantherFilenameWithRelease = releaseNumber + "_" + pantherFilename;
+        File pantherFileTar = new File(pantherFilenameWithRelease);
 
-        if (!pantherFile.exists()) {
+        if (!pantherFileTar.exists()) {
             System.out.println("Downloading " + pantherFileURL);
-            FileUtils.copyURLToFile(pantherFileURL, new File(pantherFilename));
+            FileUtils.copyURLToFile(pantherFileURL, new File(pantherFilenameWithRelease));
         } else {
-            System.out.println(pantherFile + " already exists");
+            System.out.println(pantherFileTar + " already exists");
         }
 
         JSONParser parser = new JSONParser();
         JSONObject speciesJSONFile = (JSONObject) parser.parse(new FileReader(pathToSpeciesConfig));
-        
+
         // If using an alternative source species, specify the 4-letter code as the first argument
         String sourceMappingSpecies = "";
         if (args.length > 0) {
-        	sourceMappingSpecies = args[0];
+            sourceMappingSpecies = args[0];
         } else {
-        	sourceMappingSpecies = "hsap";
+            sourceMappingSpecies = "hsap";
         }
 
-        for (Object speciesKey :  speciesJSONFile.keySet()) {
-        	
-        	JSONObject speciesJSONObject = (JSONObject) speciesJSONFile.get(speciesKey);
-        	String speciesName = (String) ((JSONArray) speciesJSONObject.get("name")).get(0);
-        	String biomartFilename = releaseNumber + "_" + speciesKey;
-        	if (speciesKey.equals(sourceMappingSpecies)) {
-        	    biomartFilename += "_protein_gene_mapping.txt";
-            } else {
-        	    biomartFilename += "_gene_protein_mapping.txt";
-            }
-        	File biomartFile = new File(biomartFilename);
+        for (Object speciesKey : speciesJSONFile.keySet()) {
 
-        	if (!biomartFile.exists()) {
+            JSONObject speciesJSONObject = (JSONObject) speciesJSONFile.get(speciesKey);
+            String speciesName = (String) ((JSONArray) speciesJSONObject.get("name")).get(0);
+            String biomartFilename = releaseNumber + "_" + speciesKey;
+            if (speciesKey.equals(sourceMappingSpecies)) {
+                biomartFilename += "_protein_gene_mapping.txt";
+            } else {
+                biomartFilename += "_gene_protein_mapping.txt";
+            }
+            File biomartFile = new File(biomartFilename);
+
+            if (!biomartFile.exists()) {
                 System.out.println("Generating Biomart mapping file for " + speciesName);
                 // Build the XML portion of the Biomart query
                 String biomartXMLQuery = BiomartUtilities.buildBiomartXML((String) speciesKey, (JSONObject) speciesJSONFile.get(speciesKey), pathToXMLQuery);
@@ -82,23 +82,35 @@ public class Main
                     MappingFileGenerator.createTargetGeneProteinMappingFile(biomartResults, (String) speciesKey, speciesName, biomartFile);
                 }
             } else {
-        	    System.out.println(biomartFile + " already exists");
+                System.out.println(biomartFile + " already exists");
             }
         }
 
-        InputStream is = new FileInputStream(pantherFile);
-        GzipCompressorInputStream gzipIn = new GzipCompressorInputStream(is);
-        TarArchiveInputStream tarIn = new TarArchiveInputStream(gzipIn);
-        TarArchiveEntry tarFile;
-        while ((tarFile = (TarArchiveEntry) tarIn.getNextEntry()) != null) {
-            int count;
-            byte data[] = new byte[1024];
-            FileOutputStream fos = new FileOutputStream(tarFile.getName(), false);
-            try (BufferedOutputStream dest = new BufferedOutputStream(fos, 1024)) {
-                while ((count = tarIn.read(data, 0, 1024)) != -1) {
-                    dest.write(data, 0, count);
+        File extractedPantherFile = new File(pantherFilename.replace(".tar.gz", ""));
+        if (!extractedPantherFile.exists()) {
+            InputStream is = new FileInputStream(pantherFileTar);
+            GzipCompressorInputStream gzipIn = new GzipCompressorInputStream(is);
+            TarArchiveInputStream tarIn = new TarArchiveInputStream(gzipIn);
+            TarArchiveEntry tarFile;
+            while ((tarFile = (TarArchiveEntry) tarIn.getNextEntry()) != null) {
+                int count;
+                byte data[] = new byte[1024];
+                FileOutputStream fos = new FileOutputStream(tarFile.getName(), false);
+                try (BufferedOutputStream dest = new BufferedOutputStream(fos, 1024)) {
+                    while ((count = tarIn.read(data, 0, 1024)) != -1) {
+                        dest.write(data, 0, count);
+                    }
                 }
+                //TODO: Version output file
             }
+        } else {
+            System.out.println("Panther file has already been extracted");
+        }
+
+        BufferedReader br = new BufferedReader(new FileReader(extractedPantherFile));
+        String line;
+        while ((line = br.readLine()) != null) {
+//            System.out.println(line);
         }
     }
 }
